+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # CHANGELOG
 
-## [1.1.4] - 2022-03-dd
+## [1.1.5] - 2022-MM-DD
+- Overload `byte` API to allow partial tracing:
+  - `Log#h(String, byte, int, int)`
+  - `getByteTraceString#(byte, int, int)`
+- Update `Application` with `Application#setInstance(android.app.Application)`
+  to allow proper operation for those which can't extend
+  `io.cloudwalk.loglibrary.Application` instead of `android.app.Application`
+  and will replace `android:name` on their `AndroidManifest.xml` anyway.
+- Review CHANGELOG.md content.
+
+## [1.1.4] - 2022-03-21
 - Replace `Application#getPackageContext()` by `Application#getContext()`.
 
 ## [1.1.3] - 2022-03-09
@@ -22,7 +32,7 @@
   `Log#h(String, byte[], int)` protection against invalid arguments.
 
 ## [1.0.1] - 2021-10-04
-- Erase dependency list to shrink final package. 
+- Erase unused dependencies from the dependency list. 
 
 ## [1.0.0] - 2021-09-30
 - Initial release.